12th march 2019

+++ cms/admin/includes/view_all_posts.php
@@ -15,6 +15,26 @@ if(isset($_POST['checkBoxArray'])){
                 $query = "DELETE FROM posts WHERE post_id = {$postValueID}";
                 mysqli_query($connection, $query);
                 break;
+            case "clone":
+                $query = "SELECT * FROM posts WHERE post_id = '{$postValueID}'";
+                $select_post_query = mysqli_query($connection, $query);
+                while($row = mysqli_fetch_array($select_post_query)){
+                    $post_title = $row["post_title"];
+                    $post_category_id = $row["post_category_id"];
+                    $post_date = $row["post_date"];
+                    $post_author = $row["post_author"];
+                    $post_status = $row["post_status"];
+                    $post_image = $row["post_image"];
+                    $post_content = $row["post_content"];
+                    $post_tag = $row["post_tag"];
+                }
+                $query = "INSERT INTO posts(post_category_id, post_title, post_author, post_date,post_image,
+              post_content, post_tag, post_status) VALUES ('{$post_category_id}',
+               '{$post_title}', '{$post_author}', now(), '{$post_image}', '{$post_content}',
+                '{$post_tags}', '{$post_status}')";
+                $create_post_query = mysqli_query($connection, $query);
+                header("Location: posts.php");
+                break;
         }
     }
 }
@@ -30,6 +50,7 @@ if(isset($_POST['checkBoxArray'])){
         <option value="published">Published</option>
         <option value="draft">Draft</option>
         <option value="delete">Delete</option>
+        <option value="clone">Clone</option>
     </select>
 </div>
 
@@ -50,6 +71,7 @@ if(isset($_POST['checkBoxArray'])){
         <th>Tags</th>
         <th>Comments</th>
         <th>Date</th>
+        <th>Views</th>
         <th>View Post</th>
         <th>Edit</th>
         <th>Delete</th>
@@ -71,6 +93,7 @@ if(isset($_POST['checkBoxArray'])){
         $post_tag = $row["post_tag"];
         $post_comment_count = $row["post_comment_count"];
         $post_date = $row["post_date"];
+        $post_view_count = $row["post_view_count"];
 
         echo "<tr>";
         ?>
@@ -97,9 +120,10 @@ if(isset($_POST['checkBoxArray'])){
         echo "<td>$post_tag</td>";
         echo "<td>$post_comment_count</td>";
         echo "<td>$post_date</td>";
+        echo "<td><a href='posts.php?reset={$post_id}'>$post_view_count</a></td>";
         echo "<td><a href='../post.php?p_id={$post_id}'>View Post</a></td>";
         echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-        echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+        echo "<td><a onclick=\"javascript : return confirm('Are you sure you want to delete this post'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
         echo "</tr>";
 
     }
@@ -109,6 +133,7 @@ if(isset($_POST['checkBoxArray'])){
 </form>
 
 <?php
+
 if (isset($_GET['delete'])) {
     $the_post_id = $_GET['delete'];
     $query = "DELETE FROM posts WHERE post_id={$the_post_id}";
@@ -116,4 +141,10 @@ if (isset($_GET['delete'])) {
     header("Location: posts.php");
 }
 
+if (isset($_GET['reset'])) {
+    $the_post_id = $_GET['reset'];
+    $query = "UPDATE posts SET post_view_count = 0 WHERE post_id={$the_post_id}";
+    $reset_post_query = mysqli_query($connection, $query);
+    header("Location: posts.php");
+}
 ?>
